config file creation reworked for all implemented problems
